+++ config/settings/base.py
@@ -477,13 +477,13 @@ IRODS_WEBDAV_URL = env.str('IRODS_WEBDAV_URL', 'https://0.0.0.0')
 
 # Irodsbackend settings
 # Status query interval in seconds
-IRODSBACKEND_STATUS_INTERVAL = env.int('IRODSBACKEND_STATUS_INTERVAL', 12)
+IRODSBACKEND_STATUS_INTERVAL = env.int('IRODSBACKEND_STATUS_INTERVAL', 15)
+# Set batch query size for improving sequential iRODS query performance (#432)
+IRODS_QUERY_BATCH_SIZE = env.int('IRODS_QUERY_BATCH_SIZE', 32)
 
 
 # Samplesheets settings
 SHEETS_IRODS_LIMIT = env.int('SHEETS_IRODS_LIMIT', 50)  # iRODS file query limit
-# Set batch query size for improving sequential iRODS query performance (#432)
-SHEETS_QUERY_BATCH_SIZE = env.int('SHEETS_QUERY_BATCH_SIZE', 16)
 SHEETS_MAX_COLUMN_WIDTH = 350  # Max column width for content overflow
 
 SHEETS_EXTERNAL_LINK_LABELS = {